% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMeanShape.R
\name{calcMeanShape}
\alias{calcMeanShape}
\title{Calculate mean shape per grid cell}
\usage{
calcMeanShape(x)
}
\arguments{
\item{x}{object of class \code{epmGrid}}
}
\value{
a list with 2 elements:
	(1) matrix where nrow = number of grid cells and ncol = 
the number of data columns. Each row is a vector of mean shape coordinates.
	(2) a matrix of xy coordinates corresponding to those grid cells.
}
\description{
For an epmGrid object that contains geometric
morphometric shape coordinates, calculate the per-grid-cell
mean shape.
}
\details{
This function will ignore cells that are empty.
}
\examples{
\donttest{
tamiasEPM
tamiasEPM <- addTraits(tamiasEPM, tamiasTraits)
meanshape <- calcMeanShape(tamiasEPM)

head(meanshape[[1]])
head(meanshape[[2]])
}
}
\author{
Pascal Title
}
