% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disparity_betadiv.R
\name{disparity_betadiv}
\alias{disparity_betadiv}
\title{Map change in morphological disparity}
\usage{
disparity_betadiv(x, radius, slow = FALSE, nThreads = 1)
}
\arguments{
\item{x}{object of class \code{epmGrid}.}

\item{radius}{Radius of the moving window in map units.}

\item{slow}{if TRUE, use an alternate implementation that has a lower memory footprint 
but that is likely to be much slower. Most useful for high spatial resolution.}

\item{nThreads}{number of threads for parallelization}
}
\value{
Returns a sf polygons object with mean partial disparity.
}
\description{
Change in morphological partial disparity is calculating across a moving window
of neighboring grid cells.
}
\details{
Disparity is calculated for each cell + its neighboring cells. Partial disparity is then calculated
	for each species, and mean partial disparity of species in neighboring cells (but not the focal cell)
	is calculated and assigned to the focal cell. The value at each grid cell is therefore a measure of how 
	much neighbors contribute to the overall local disparity.
}
\examples{
\donttest{
tamiasEPM

tamiasEPM <- addTraits(tamiasEPM, tamiasTraits)

z <- disparity_betadiv(tamiasEPM, radius = 150000)

plot(z)

# using square grid epmGrid
tamiasEPM2 <- createEPMgrid(tamiasPolyList, resolution = 50000, 
	cellType = 'square', method = 'centroid')
tamiasEPM2 <- addTraits(tamiasEPM2, tamiasTraits)
z2 <- disparity_betadiv(tamiasEPM2, radius = 150000)

terra::plot(z2, col = sf::sf.colors(100), range = c(0,1))

}
}
\references{
Foote M. 1993. Contributions of individual taxa to overall morphological disparity. Paleobiology. 19:403–419.
}
\author{
Pascal Title
}
