% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRstat.R
\name{DRstat}
\alias{DRstat}
\title{Calculate the DR statistic}
\usage{
DRstat(tree)
}
\arguments{
\item{tree}{phylogeny of class \code{phylo}}
}
\value{
named numeric vector of speciation rates
}
\description{
Calculates the tip-specific DR statistic for speciation rates
}
\examples{
tamiasTree
DRstat(tamiasTree)

}
\references{
Jetz, W., Thomas, G. H., Joy, J. B., Hartmann, K., & Mooers, A. O. (2012). 
	The global diversity of birds in space and time. Nature, 491, 444–448.
}
\author{
Pascal Title
}
