% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epm-example}
\alias{epm-example}
\alias{tamiasEPM}
\alias{tamiasPolyList}
\alias{tamiasTraits}
\alias{tamiasTree}
\title{Eco Phylo Mapper datasets}
\format{
An object of class \code{epmGrid} of length 7.
}
\usage{
tamiasEPM
}
\description{
Included datasets in \code{epm}
}
\details{
Included north american chipmunk dataset: \cr
tamiasTree is a phylogeny for chipmunks from Zelditch et al. 2017 \cr
tamiasTraits is a geometric morphometrics dataset of mean values for 
chipmunks from Zelditch et al. 2017 \cr
tamiasPolyList is a set of geographic ranges for chimpunks from IUCN 2021.\cr
tamiasEPM is an \code{epmGrid} object created with \code{\link{createEPMgrid}} 
using these datasets.
}
\references{
Zelditch, M. L., Ye, J., Mitchell, J. S., & Swiderski, D. L.
(2017). Rare ecomorphological convergence on a complex adaptive landscape:
Body size and diet mediate evolution of jaw shape in squirrels (Sciuridae).
Evolution, 1–17. https://doi.org/10.1111/evo.13168 
IUCN 2021. The IUCN Red List of Threatened Species. 2021-3. 
\url{https://www.iucnredlist.org}. 'Downloaded on 17 March 2021.
}
\keyword{datasets}
