% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExtentOfList.R
\name{getExtentOfList}
\alias{getExtentOfList}
\title{Get extent of list}
\usage{
getExtentOfList(shapes, format = "raster")
}
\arguments{
\item{shapes}{a list of SpatialPolygons or simple features}

\item{format}{if \code{raster}, then returns a raster extent object,
otherwise, returns a sf formatted bbox object.}
}
\value{
An object of class \code{extent} or \code{bbox}.
}
\description{
Given a list of SpatialPolygons, return an extent
object that encompasses all items.
}
\examples{
getExtentOfList(tamiasPolyList)

}
\author{
Pascal Title
}
