% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.epmGrid.R
\name{identify.epmGrid}
\alias{identify.epmGrid}
\title{Interactively identify species in epmGrid}
\usage{
\method{identify}{epmGrid}(x, returnCell = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{epmGrid} or \code{sf}}

\item{returnCell}{boolean; if FALSE, then species names are returned,
if TRUE, then cell indices are returned.}

\item{...}{additional arguments passed to sf::plot}
}
\value{
A vector of species names or cell indices.
}
\description{
Plots a epmGrid object and allows you to click on the plot
	to return the species found in the cell you clicked on.
}
\details{
This is a wrapper function for the \code{identify} function
in base graphics. This is primarily intended as a useful function for 
data exploration and spot-checking.
}
\author{
Pascal Title
}
