% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveExtent.R
\name{interactiveExtent}
\alias{interactiveExtent}
\title{Interactively choose extent}
\usage{
interactiveExtent(polyList, cellType = "square", bb = NULL, nThreads = 1)
}
\arguments{
\item{polyList}{a list of Simple Feature polygons.}

\item{cellType}{either \code{hexagon} or \code{square}.}

\item{bb}{c(xmin, xmax, ymin, ymax) to limit the extent for 
the interactive plot.}

\item{nThreads}{if > 1, then employ parallel computing.}
}
\value{
A list with a polygon, and its WKT string
}
\description{
Given a list of polygons, sets up an interactive plot
	to allow the user to draw the desired extent. This can be used to
	define the extent in \link{createEPMgrid}.
}
\examples{
\dontrun{
ex <- interactiveExtent(tamiasPolyList)

# You can use this as the extent in createEPMgrid
grid <- createEPMgrid(tamiasPolyList, resolution = 50000, extent = ex$wkt)

}
}
\author{
Pascal Title
}
