% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic_betadiv.R
\name{phylogenetic_betadiv}
\alias{phylogenetic_betadiv}
\title{Map turnover in species communities}
\usage{
phylogenetic_betadiv(x, radius, component = "full", slow = FALSE, nThreads = 1)
}
\arguments{
\item{x}{object of class \code{epmGrid}.}

\item{radius}{Radius of the moving window in map units.}

\item{component}{which component of beta diversity to use, can be \code{"turnover"}, 
\code{"nestedness"} or \code{"full"}}

\item{slow}{if TRUE, use an alternate implementation that has a lower memory footprint 
but that is likely to be much slower. Most useful for high spatial resolution.}

\item{nThreads}{number of threads for parallelization}
}
\value{
Returns a sf polygons object with mean community dissimilarity for each grid cell.
}
\description{
Mean phylogenetic community dissimilarity is 
	calculated for each cell within a circular moving window of neighboring cells.
}
\details{
For each cell, mean dissimilarity is calculated from the focal cell to each of its neighbors.

	All metrics are based on Sorensen dissimilarity and range from 0 to 1:
	For each metric, the following components can be specified. These components are additive,
	such that the full metric = turnover + nestedness. 
	\itemize{
		\item{turnover}: species turnover without the influence of richness differences
		\item{nestedness}: species turnover due to differences in richness
		\item{full}: the combined turnover due to both differences in richness and pure turnover
	}
}
\examples{
\donttest{
tamiasEPM

tamiasEPM <- addPhylo(tamiasEPM, tamiasTree)

# phylogenetic turnover
beta_phylo_turnover <- phylogenetic_betadiv(tamiasEPM, radius = 70000, 
		component = 'turnover')
beta_phylo_nestedness <- phylogenetic_betadiv(tamiasEPM, radius = 70000, 
		component = 'nestedness')
beta_phylo_full <- phylogenetic_betadiv(tamiasEPM, radius = 70000, 
		component = 'full')


colramp <- colorRampPalette(c('blue','yellow','red'))
par(mfrow=c(1,3))
plot(beta_phylo_turnover, col = colramp(100))
plot(beta_phylo_nestedness, col = colramp(100))
plot(beta_phylo_full, col = colramp(100))

}
}
\references{
Baselga, A. The relationship between species replacement, dissimilarity derived from nestedness, 
and nestedness. Global Ecology and Biogeography 21 (2012): 1223–1232.

Laffan, SW, et al. Range-weighted metrics of species and phylogenetic turnover can better 
resolve biogeographic transition zones. Methods in Ecology and Evolution 7 (2016): 580-588.

Leprieur, F, Albouy, C, De Bortoli, J, Cowman, PF, Bellwood, DR & Mouillot, D. Quantifying 
Phylogenetic Beta Diversity: Distinguishing between "True" Turnover of Lineages and Phylogenetic 
Diversity Gradients. PLoS ONE 7 (2012): e42760–12.

Rosauer, D, Laffan, SW, Crisp, MD, Donnellan, SC, Cook, LG. Phylogenetic endemism: a new approach 
for identifying geographical concentrations of evolutionary history. Molecular Ecology 
18 (2009): 4061-4072.
}
\author{
Pascal Title
}
