% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.epmGrid.R
\name{read.epmGrid}
\alias{read.epmGrid}
\title{Read a epmGrid object}
\usage{
read.epmGrid(filename)
}
\arguments{
\item{filename}{filename, with extension \code{rds}}
}
\value{
object of class epmGrid
}
\description{
Load a saved epmGrid object.
}
\details{
This function will read in epmGrid objects that
were saved with \code{\link{write.epmGrid}}.
}
\examples{
\dontrun{
#save
write.epmGrid(tamiasEPM, '~/tamiasEPM')

# read back in
tamiasEPM <- read.epmGrid('~/tamiasEPM.rds')
}
}
\author{
Pascal Title
}
