% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuildEpmGrid.R
\name{rebuildEpmGrid}
\alias{rebuildEpmGrid}
\title{Rebuild epmGrid}
\usage{
rebuildEpmGrid(x, spCellList)
}
\arguments{
\item{x}{object of class \code{epmGrid}}

\item{spCellList}{list in which each element is a character vector of species names.}
}
\value{
new \code{epmGrid} object.
}
\description{
Given a list of species in each cell, all components of the
	epmGrid object are reindexed and regenerated.
}
\details{
This function is used internally by \code{\link{dropSpecies}}.
}
\author{
Pascal Title
}
