% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSpCellIndex.R
\name{singleSpCellIndex}
\alias{singleSpCellIndex}
\title{Identify single-species cells}
\usage{
singleSpCellIndex(x)
}
\arguments{
\item{x}{object of class \code{epmGrid}}
}
\value{
numeric vector of grid cell indices.
}
\description{
Given a epmGrid object, return the grid cell indices
	of those cells that have just one species.
}
\details{
This function can be useful when further analyzing epmGrid
	objects generated by \code{\link{gridMetrics}}, as it might
	make sense to exclude these single-species cells in further analyses.
}
\examples{
singleSpCellIndex(tamiasEPM)

}
\author{
Pascal Title
}
