% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.EPMgrid.R
\name{summary.epmGrid}
\alias{summary.epmGrid}
\title{epmGrid summary}
\usage{
\method{summary}{epmGrid}(object, ...)
}
\arguments{
\item{object}{object of class \code{epmGrid}}

\item{...}{further arguments passed to \code{\link{summary}}}
}
\value{
A list containing the summary information is returned invisibly.
}
\description{
Generates a summary of a epmGrid object.
}
\details{
Summary information includes
}
\examples{
summary(tamiasEPM)
attr <- summary(tamiasEPM)
attr

}
\author{
Pascal Title
}
