% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEPMgrid.R
\name{createEPMgrid}
\alias{createEPMgrid}
\title{Create epmGrid object}
\usage{
createEPMgrid(
  polyList,
  resolution = 50000,
  method = "centroid",
  cellType = "hexagon",
  coverCutoff = 0.1,
  retainSmallRanges = TRUE,
  extent = "auto",
  checkValidity = FALSE,
  nThreads = 1,
  nGroups = 1
)
}
\arguments{
\item{polyList}{a list of polygon objects (sf or sp), named with taxon names.}

\item{resolution}{vertical and horizontal spacing of grid cells, in units 
of the polygons' projection.}

\item{method}{approach used for gridding. Either \code{centroid} or \code{areaCutoff}. See details below.}

\item{cellType}{either \code{hexagon} or \code{square}. See details below.}

\item{coverCutoff}{the percent that a species range must cover a grid cell to be considered present.}

\item{retainSmallRanges}{boolean; should small ranged species be dropped or preserved.
See details.}

\item{extent}{if 'auto', then the maximal extent of the polygons will be used. 
If not 'auto', can be a SpatialPolygon or sf object, in which case the resulting epmGrid
will be cropped and masked with respect to the polygon, or a spatial coordinates object, 
from which an extent object will be generated, or a numeric vector of length 4 
with minLong, maxLong, minLat, maxLat. If 'interactive', then an interactive plot
will appear in which the user can draw the desired polygon extent. That extent will then be returned.}

\item{checkValidity}{if \code{TRUE}, then check polygon validity and repair if needed, 
using sf::st_make_valid and the lwgeom package.}

\item{nThreads}{if > 1, then employ parallel computing.}

\item{nGroups}{break up the grid into this many groups for processing. This will alleviate memory 
usage for datasets that are very high resolution (not relevant for large numbers of species)}
}
\value{
an object of class \code{epmGrid}. If \code{extent = 'interactive'}, then a polygon is returned.
}
\description{
Takes a list of polygons and creates a epmGrid object.
}
\details{
If \code{cellType = 'hexagon'}, then the grid is made of polygons via the sf package.
	If \code{cellType = 'square'}, then the grid is a raster generated via the terra package.
	Hexagonal cells have several advantages, including being able to be of different sizes (if the grid is in 
	unprojected long/lat), and may be able to more naturally follow coastlines and non-linear features.
	However, the raster-based square cells will be much less memory intensive for high resolution datasets. 
	Choice of grid type matters more for spatial resolution, than for number of species.


	In the polygon-to-grid conversion process, two approaches are implemented. 
	For \code{method = 'centroid'}, a range polygon registers in a cell if the polygon 
	overlaps with the cell centroid. 
	For \code{method = 'areaCutoff'}, a range polygon registers in a cell if it covers that cell by 
	at least \code{coverCutoff} fraction of the cell.
	If \code{retainSmallRanges = FALSE}, then species whose ranges are so small that no 
	cell registers as present will be dropped. If \code{retainSmallRanges = TRUE}, then the 
	cell that contains the majority of the the small polygon will be considered as present, even if it's a 
	small percent of the cell.

	If \code{retainSmallRanges = TRUE}, and an extent is provided, then species may still be dropped if they
	fall outside of that extent.

	In interactive mode for defining the extent, the user can draw a bounding polygon on a 
	map. The drawn polygon will then be printed to the console so that the user can provide 
	that bounding polygon in future calls as the extent.

	Any SpatialPolygon or SpatialPoints objects are converted to objects of class \code{sf}.


	In interactive mode, the basemap is from \url{www.naturalearthdata.com}.
}
\examples{
library(sf)
# example dataset: a list of 24 chipmunk distributions as polygons
head(tamiasPolyList)

# hexagonal grid
tamiasEPM <- createEPMgrid(tamiasPolyList, resolution = 50000, 
	cellType = 'hexagon', method = 'centroid')
tamiasEPM

# square grid
tamiasEPM2 <- createEPMgrid(tamiasPolyList, resolution = 50000, 
	cellType = 'square', method = 'centroid')
tamiasEPM2

}
\author{
Pascal Title
}
