% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{epm-example}
\alias{epm-example}
\alias{tamiasEPM}
\alias{tamiasPolyList}
\alias{tamiasTraits}
\alias{tamiasTree}
\title{Eco Phylo Mapper datasets}
\description{
Included datasets in \code{epm}
}
\details{
Included north american chipmunk dataset
}
