% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm-package.R
\docType{package}
\name{epm}
\alias{epm}
\title{EcoPhyloMapper (epm)}
\description{
An R package that facilitates the aggregation of species'
geographic ranges from vector or raster spatial data, and that enables
the calculation of various morphological and phylogenetic metacommunity metrics
across geography.

A detailed wiki for the R package can be found on the epm github page:
\url{https://github.com/ptitle/epm/wiki#table-of-contents}
}
\author{
Pascal O. Title, Donald L. Swiderski, Miriam L. Zelditch
}
