% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridMetrics.R
\name{gridMetrics}
\alias{gridMetrics}
\title{Grid Metrics}
\usage{
gridMetrics(x, metric, var = NULL, nreps = 20, taxa = NULL, verbose = FALSE)
}
\arguments{
\item{x}{object of class \code{epmGrid}}

\item{metric}{name of metric to use, see Details.}

\item{var}{If a univariate morphological metric is specified, and the 
data in \code{x} are multivariate, which trait should be used?
This can also specify which subset of columns a multivariate metric should be applied to.}

\item{nreps}{Number of repetitions for Foote metric distribution.}

\item{verbose}{Intended primarily for debugging, prints progress to the console}
}
\value{
object of class \code{epmGrid} where the grid represents
	calculations of the metric at every cell.
}
\description{
Calculate various morphological and phylogenetic community
metrics for every cell in a \code{epmGrid} object.
}
\details{
Univariate trait metrics
	\itemize{
		\item{mean}
		\item{median}
		\item{range}
		\item{mean_NN_dist:} {mean nearest neighbor distance}
		\item{min_NN_dist:} {minimum nearest neighbor distance}
		\item{variance}
		\item{arithmeticWeightedMean} (see below)
		\item{geometricWeightedMean} (see below)
		\item{phylosignal:} {Blomberg's K for phylogenetic signal}
	}
	Multivariate trait metrics
	\itemize{
		\item{disparity} 
		\item{partialDisparity:} {contribution of species in each gridcell to overall disparity, returned as the ratio of summed partial disparities to total disparity.}
		\item{range}
		\item{rangePCA}
		\item{mean_NN_dist:} {mean nearest neighbor distance}
		\item{min_NN_dist:} {minimum nearest neighbor distance}
		\item{phylosignal:} {Blomberg's K for phylogenetic signal, as implemented in geomorph::physignal}
	}
	Phylogenetic metrics
	\itemize{
		\item{pd:} {Faith's phylogenetic diversity, including the root}
		\item{meanPatristic}
		\item{meanPatristicNN:} {mean nearest neighbor in patristic distance}
		\item{minPatristicNN:} {minimum nearest neighbor in patristic distance}
		\item{phyloDisparity:} {sum of squared deviations in patristic distance}
		\item{PSV:} {Phylogenetic Species Variability}
		\item{PSR:} P={Phylogenetic Species Richness}
		\item{DR:} {non-parametric estimate of speciation rates}
	}
	Range-weighted metrics
	\itemize{
		\item{weightedEndemism}
		\item{correctedWeightedEndemism:} {Weighted endemism standardized by species richness}
		\item{phyloWeightedEndemism:}
	}

	If data slot contains a pairwise matrix, \code{var} is ignored.
	Weighted mean options are available where, for each cell, a weighting scheme 
	(inverse of species range sizes) is applied such that small-ranged species are 
	up-weighted, and broadly distributed species are down-weighted. 
	This can be a useful way to lessen the influence of broadly distributed species 
	in the geographic mapping of trait data.
}
\examples{
tamiasEPM <- addPhylo(tamiasEPM, tamiasTree)
tamiasEPM <- addTraits(tamiasEPM, tamiasTraits)

# univariate morphological example
x <- gridMetrics(tamiasEPM, metric='mean', var='V2')
plot(x) 

# multivariate morphological
x <- gridMetrics(tamiasEPM, metric='disparity')
plot(x)

# phylogenetic metrics
x <- gridMetrics(tamiasEPM, metric='meanPatristic')
plot(x)

}
