% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSpeciesRaster.R
\name{write.epmGrid}
\alias{write.epmGrid}
\title{Save epmGrid object}
\usage{
write.epmGrid(x, filename)
}
\arguments{
\item{x}{object of class \code{epmGrid}}

\item{filename}{filename with no extension}
}
\value{
Nothing is returned, but object is written to disk.
}
\description{
Write a epmGrid object to disk.
}
\details{
This function writes a .rds file with xz compression. 
This file can be read back in with \code{\link{readRDS}}.
}
\examples{
\dontrun{
#save
save(tamiasEPM, '~/tamiasEPM')

# read back in
tamiasEPM <- readRDS('~/tamiasEPM.rds')
}
}
\author{
Pascal Title
}
